pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ticketflow-api"
version = "1.0.0"
description = "Sistema didático de suporte técnico em tempo real com FastAPI, Pydantic v2, Beanie e MongoDB."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Romario" },
]
keywords = ["fastapi", "mongodb", "beanie", "sse", "websocket", "webhook"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Education",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]

dependencies = [
    "fastapi[standard]==0.115.12",
    "uvicorn[standard]==0.34.0",
    "beanie==1.29.0",
    "motor==3.7.0",
    "pydantic[email]==2.11.1",
    "pydantic-settings==2.8.1",
    "python-jose[cryptography]==3.4.0",
    "pwdlib[argon2]==0.3.0",
    "python-multipart==0.0.20",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.fastapi]
entrypoint = "app.main:app"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
